# openapi.json
{"openapi": "3.1.0", "info": {"title": "Alpha Geometry API", "description": "\n                    API for running alphageometry in different configurations.\n                  ", "version": "v1"}, "paths": {"/healthcheck": {"get": {"summary": "Perform Healthcheck", "operationId": "perform_healthcheck_healthcheck_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/run_ddar": {"post": {"summary": "Run Ddar", "operationId": "run_ddar_run_ddar_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConstructionStr"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Run Ddar Run Ddar Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/plot": {"post": {"summary": "Plot Figure", "operationId": "plot_figure_plot_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConstructionStr"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"ConstructionStr": {"properties": {"txt": {"type": "string", "title": "Txt"}}, "type": "object", "required": ["txt"], "title": "ConstructionStr"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}